// src/api/models/business-user-edition-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { BusinessUserTypeSaveModel } from './business-user-type-save-model';
// May contain unused imports in some cases
// @ts-ignore
import { Status } from './status';

/**
 * 
 * @export
 * @interface BusinessUserEditionModel
 */
export interface BusinessUserEditionModel {
    /**
     * 
     * @type {BusinessUserTypeSaveModel}
     * @memberof BusinessUserEditionModel
     */
    'type'?: BusinessUserTypeSaveModel;
    /**
     * 
     * @type {number}
     * @memberof BusinessUserEditionModel
     */
    'responsibleBusinessUserId'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof BusinessUserEditionModel
     */
    'id'?: number;
    /**
     * 
     * @type {Status}
     * @memberof BusinessUserEditionModel
     */
    'status'?: Status;
}