// src/api/models/business-user-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { BusinessModel } from './business-model';
// May contain unused imports in some cases
// @ts-ignore
import { BusinessUserDataByTypeModel } from './business-user-data-by-type-model';
// May contain unused imports in some cases
// @ts-ignore
import { Status } from './status';
// May contain unused imports in some cases
// @ts-ignore
import { UserModel } from './user-model';

/**
 * 
 * @export
 * @interface BusinessUserModel
 */
export interface BusinessUserModel {
    /**
     * 
     * @type {number}
     * @memberof BusinessUserModel
     */
    'id'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof BusinessUserModel
     */
    'isNullable'?: boolean;
    /**
     * 
     * @type {Status}
     * @memberof BusinessUserModel
     */
    'status'?: Status;
    /**
     * 
     * @type {string}
     * @memberof BusinessUserModel
     */
    'createdBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessUserModel
     */
    'createdOnUtc'?: string;
    /**
     * 
     * @type {string}
     * @memberof BusinessUserModel
     */
    'updatedBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessUserModel
     */
    'updatedOnUtc'?: string | null;
    /**
     * 
     * @type {BusinessUserDataByTypeModel}
     * @memberof BusinessUserModel
     */
    'type'?: BusinessUserDataByTypeModel;
    /**
     * 
     * @type {number}
     * @memberof BusinessUserModel
     */
    'businessId'?: number;
    /**
     * 
     * @type {BusinessModel}
     * @memberof BusinessUserModel
     */
    'business'?: BusinessModel;
    /**
     * 
     * @type {string}
     * @memberof BusinessUserModel
     */
    'userId'?: string | null;
    /**
     * 
     * @type {UserModel}
     * @memberof BusinessUserModel
     */
    'user'?: UserModel;
    /**
     * 
     * @type {number}
     * @memberof BusinessUserModel
     */
    'responsibleBusinessUserId'?: number | null;
    /**
     * 
     * @type {BusinessUserModel}
     * @memberof BusinessUserModel
     */
    'responsibleBusinessUser'?: BusinessUserModel;
}