// src/api/models/user-creation-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { BusinessUserTypeSaveModel } from './business-user-type-save-model';

/**
 * 
 * @export
 * @interface UserCreationModel
 */
export interface UserCreationModel {
    /**
     * 
     * @type {string}
     * @memberof UserCreationModel
     */
    'firstName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserCreationModel
     */
    'secondName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserCreationModel
     */
    'firstLastName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserCreationModel
     */
    'secondLastName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserCreationModel
     */
    'phoneNumber'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserCreationModel
     */
    'userName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserCreationModel
     */
    'password'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserCreationModel
     */
    'email'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof UserCreationModel
     */
    'businessId'?: number | null;
    /**
     * 
     * @type {BusinessUserTypeSaveModel}
     * @memberof UserCreationModel
     */
    'businessType'?: BusinessUserTypeSaveModel;
}